refactor: move sections into individual files
